Don't require a GitHub access token if in "skip publish" mode
The `skip-publish` option allows you to build your Gatsby site
but not publish it, "effectively performing a test of the build
process using the live configuration". This is useful for CI
purposes - i.e. checking your site builds, but not actually
deploying.

However, at the moment, you have to provide an `access-token` when
in "skip publish" mode, even though this is only used to publish.

This updates the action - including new tests - so that you are
not required to specify the `access-token` option if you've asked
to skip publishing.